--- lib/tictactoe/version.rb
@@ -1,3 +1,0 @@
-module TicTacToe
-  VERSION = "0.1.0"
-end
