+++ lib/tictactoe_core/version.rb
@@ -1,0 +1,3 @@
+module TicTacToeCore
+  VERSION = "0.1.0"
+end
