--- spec/tictactoe/fakes/game.rb
@@ -1,9 +1,0 @@
-module TicTacToe
-  module Fakes
-    class Game
-      def run
-
-      end
-    end
-  end
-end
