+++ spec/tictactoe_core/board_helper.rb
@@ -1,0 +1,14 @@
+require 'spec_helper'
+require 'tictactoe_core/board'
+
+module TicTacToeCore
+  class BoardHelper
+    def self.create_initial_board_three(layout)
+      Board.make_board(3, layout.split(''))
+    end
+  
+    def self.create_initial_board_four(layout)
+      Board.make_board(4, layout.split(''))
+    end
+  end
+end
