+++ spec/tictactoe_core/fakes/game.rb
@@ -1,0 +1,9 @@
+module TicTacToeCore
+  module Fakes
+    class Game
+      def run
+
+      end
+    end
+  end
+end
