+++ spec/tictactoe_core/game_spec.rb
@@ -1,0 +1,106 @@
+require 'spec_helper'
+require 'tictactoe_core/game'
+require 'tictactoe_core/fakes/ui_mock'
+require 'tictactoe_core/board_helper'
+
+module TicTacToeCore
+  describe Game do
+    let(:player_1) { PlayerSpy.new }
+    let(:player_2) { PlayerSpy.new }
+    let(:players) { [player_1, player_2] }
+    let(:ui) { Fakes::UiMock.new }
+
+    def initial_board(layout)
+      BoardHelper.create_initial_board_three(layout)
+    end
+
+    it "is running when you have a new game" do
+      board = Board.make_board(3)
+      game = Game.new(board, players, ui)
+      expect(game.running?).to be true
+    end
+
+    it "is not running when the board is full" do
+      board = initial_board('XXXOOOXXX')
+      game = Game.new(board, players, ui)
+      expect(game.running?).to be false
+    end
+
+    it "update will make a move on the board" do
+      board = Board.make_board(3)
+      game = Game.new(board, players, ui)
+      game.update
+      expect(game.running?).to be true
+    end
+
+    it "sets up 2 player and expect player 1 to make a move after one update" do
+      board = Board.make_board(3)
+      game = Game.new(board, players, ui)
+      game.update
+
+      expect(player_1.next_move_times_called).to eq(1)
+      expect(player_2.next_move_times_called).to eq(0)
+    end
+
+    it "changes the player between each move" do
+      board = Board.make_board(3)
+      game = Game.new(board, players, ui)
+      game.update
+      game.update
+      game.update
+
+      expect(player_1.next_move_times_called).to eq(2)
+      expect(player_2.next_move_times_called).to eq(1)
+    end
+
+    it "draws the board" do
+      board = Board.make_board(3)
+      game = Game.new(board, players, ui)
+
+      game.draw
+      expect(ui.clear_screen_times_called).to eq(1)
+      expect(ui.draw_board_times_called).to eq(1)
+    end
+
+    it "displays the game over message if the game is no longer running" do
+      board = initial_board('XXXOOOXXX')
+      game = Game.new(board, players, ui)
+
+      game.draw
+      expect(ui.display_end_game_message_times_called).to eq(1)
+    end
+
+    it "returns the game is a draw correctly" do
+      board = initial_board('OXOOXOXOX')
+      game = Game.new(board, players, ui)
+
+      expect(game.end_game_state).to eq(Board::DRAW)
+    end
+
+    it "exits the run method once the game is over" do
+      board = initial_board('XOXOXOOX-')
+      game = Game.new(board, players, ui)
+      game.run
+    end
+
+    it "checks current player marker is X at the start of the game" do
+      board = Board.make_board(3)
+      game = Game.new(board, players, ui)
+      expect(game.current_player_marker).to eq(Marker::X_MARKER)
+    end
+  end
+
+  class PlayerSpy
+    attr_reader :next_move_times_called, :marker
+
+    def initialize
+      @marker = 'X'
+      @next_move_times_called = 0
+    end
+
+    def next_move(board)
+      @next_move_times_called += 1
+      board.available_moves.sample
+    end
+  end
+end
